test: files organisation modification
