+++ test/scenes_basic.h
@@ -1,10 +1,5 @@
-#ifndef TEST_SCENES_H
-#define TEST_SCENES_H
-
-#include "node.h"
-#include "geometry/geometry.h"
-
-void new_geom_nxn(struct Geometry* geom, int len, float space, struct Node* root);
+#ifndef TEST_SCENES_BASIC_H
+#define TEST_SCENES_BASIC_H
 
 struct Node *box_10x10(struct Geometry* mat, float size, float spacing);
