+++ test/scenes_util.h
@@ -1,0 +1,9 @@
+#ifndef TEST_SCENES_UTIL_H
+#define TEST_SCENES_UTIL_H
+
+#include "node.h"
+#include "geometry/geometry.h"
+
+void new_geom_nxn(struct Geometry* geom, int len, float space, struct Node* root);
+
+#endif
